Put LND things in lnd.go and alby things in alby.go
